+++ src/SafeOrbit/CHANGELOG.md
@@ -11,13 +11,17 @@
   - Most of the public methods are now only async: if it can be async, it's async.
   - Adding byte to SafeBytes is now 2100% faster.
     - Adding multiple bytes with AppendMany is 2100% * byte amount faster.
-    - SafeString uses SafeBytes internally, and now it's also 2100% faster.
+    - SafeString uses SafeBytes internally, and now it's also 8200% faster per character.
   - `GetHashCode()` in `SafeString` and `SafeBytes` is faster than light (0 ms) compared to older implementation (6s for 1000 char with a linear increase per char)
     - The method is also used in equality checks which makes .Equals() much faster.
   - Decrypting all bytes in `SafeBytes` is now 214x faster for 10KB, much faster for bigger sizes (1 MB = 900 ms).
   - `AppendAsync(ISafeBytes)` and `DeepCloneAsync` are optimized in `SafeBytes`
     - It is much faster (from 170 seconds to 37 ms for 10KB data)
     - Leads to much faster ToSafeBytes in SafeString
+  - Faster equality checks for bytes and strings without compramising security.
+	- `SafeBytes.EqualsAsync(bytes)` for 100 KB; before: 93032ms, after: 20ms
+	- `SafeBytes.EqualsAsync(ISafeBytes)` for 100 KB before: 183346ms, after: 15ms
+    - `SafeString.EqualsAsync` gains huge performance from inner byte equality, also it does not encrypt & decrypt plain strings anymore.
 - **Added**
   - Encryption
     - Support for padding in Blowfish algorithm.